pertemuan4

+++ src/tugas/pertemuan4/Tugas/Node.java
@@ -1,15 +1,15 @@
 package pertemuan4.Tugas;
 
 public class Node {
-    private double data;
+    private int data;
     private Node next;
 
-    public Node(double data) {
+    public Node(int data) {
         this.data = data;
         this.next = null;
     }
 
-    public double getData() {
+    public int getData() {
         return data;
     }
 
@@ -22,3 +22,28 @@ public class Node {
     }
 
 }
+
+
+// jika double gunakan :
+
+//public class Node {
+//    private double data;
+//    private Node next;
+//
+//    public Node(double data) {
+//        this.data = data;
+//        this.next = null;
+//    }
+//
+//    public double getData() {
+//        return data;
+//    }
+//
+//    public Node getNext() {
+//        return next;
+//    }
+//
+//    public void setNext(Node next) {
+//        this.next = next;
+//    }
+//}
